+++ oct2py/tests/roundtrip.m
@@ -3,11 +3,11 @@ function x, class = roundtrip(y)
 
   % returns the variable it was given, and optionally the class
 
-  x = y
+  x = y;
 
   if nargout == 2
 
-	 class = class(x)
+	 class = class(x);
 
   end