+++ HISTORY.rst
@@ -3,17 +3,27 @@
 Release History
 ---------------
 
-1.2.0
+1.3.0 (2014-01-12)
+++++++++++++++++++
+- Added support for Octave keyboard function (requires pexpect on Linux).
+- (Linux) When pexpect is installed, Octave no longer closes session when
+  a Syntax Error is encountered.
+- Fixed: M-files with no docstrings are now supported.
+
+
+1.2.0 (2013-12-14)
 ++++++++++++++++++
 - OctaveMagic is now part of Oct2Py: ``%load_ext oct2py.ipython``
 - Enhanced Struct behavior - supports REPL completion and pickling
 - Fixed: Oct2Py will install on Python3 when using setup.py
+
 
 1.1.1 (2013-11-14)
 ++++++++++++++++++
 - Added support for wheels.
 - Fixed: Put docs back in the manifest.
 - Fixed: Oct2py will install when there is no Octave available.
+
 
 1.1.0 (2013-10-27)
 ++++++++++++++++++